+++ safesafe/Components/Debug/DebugViewController.swift
@@ -82,6 +82,7 @@ final class DebugViewController: ViewController<DebugViewModel> {
         stackButton(DebugViewModel.Texts.shareUploadedPayloadsTitle, action: .uploadedPayloadsShare)
         stackButton(DebugViewModel.Texts.shareLogsTitle, action: .logsShare)
         stackButton(DebugViewModel.Texts.dumpLocalStorageTitl, action: .dumpLocalstorage)
+        stackButton(DebugViewModel.Texts.downloadDistrictsTitle, action: .downloadDistricts)
     }
     
     private func stackButton(_ title: String, action: DebugAction) {
@@ -138,8 +139,21 @@ extension DebugViewController: DebugViewModelDelegate {
     }
     
     func shareLogs(fileURL: URL) {
-        let activityController = UIActivityViewController(activityItems: [fileURL], applicationActivities: nil)
-        present(activityController, animated: true)
+        let alertController = UIAlertController(title: "Pick Action", message: nil, preferredStyle: .actionSheet)
+        let preview = UIAlertAction(title: "Preview", style: .default) { [weak self] _ in
+            guard let data = try? Data(contentsOf: fileURL), let text = String(bytes: data, encoding: .utf8) else { return }
+            self?.showTextPreview(text: text)
+        }
+        let share = UIAlertAction(title: "Share", style: .default) { [weak self] _ in
+            let activityController = UIActivityViewController(activityItems: [fileURL], applicationActivities: nil)
+            self?.present(activityController, animated: true)
+        }
+        
+        alertController.addAction(preview)
+        alertController.addAction(share)
+        
+        present(alertController, animated: true)
+        
     }
     
     func showTextPreview(text: String) {