+++ safesafe/Services/Notifications/NotificationManager.swift
@@ -13,19 +13,28 @@ import FirebaseMessaging
 import PromiseKit
 
 protocol NotificationManagerProtocol {
-    func registerForRemoteNotifications() -> Guarantee<Bool>
+    func registerAPNSIfNeeded()
+    func registerForNotifications(remote: Bool) -> Guarantee<Bool>
     func currentStatus() -> Guarantee<UNAuthorizationStatus>
     func clearBadgeNumber()
     func update(token: Data)
     func unsubscribeFromDailyTopic(timestamp: TimeInterval)
     func stringifyUserInfo() -> String?
     func clearUserInfo()
+    func showDistrictStatusLocalNotification(with changed: [DistrictStorageModel], observed: [ObservedDistrictStorageModel], timestamp: Int, delay: TimeInterval)
+}
+
+extension NotificationManagerProtocol {
+    func registerForNotifications(remote: Bool = true) -> Guarantee<Bool> {
+        return registerForNotifications(remote: remote)
+    }
 }
 
 final class NotificationManager: NSObject {
     
     enum Constants {
         static let dailyTopicDateFormat = "ddMMyyyy"
+        static let districtNotificationIdentifier = "DistrictNotificationID"
     }
     
     static let shared = NotificationManager()
@@ -111,7 +120,7 @@ extension NotificationManager: NotificationManagerProtocol {
         }
     }
     
-    func registerForRemoteNotifications() -> Guarantee<Bool> {
+    func registerForNotifications(remote: Bool = true) -> Guarantee<Bool> {
         return Guarantee { fulfill in
             let didRegister = StoredDefaults.standard.get(key: .didAuthorizeAPN) ?? false
             guard !didRegister else {
@@ -125,14 +134,49 @@ extension NotificationManager: NotificationManagerProtocol {
                     return
                 }
                 
-                DispatchQueue.main.async {
-                    UIApplication.shared.registerForRemoteNotifications()
+                if remote {
+                    DispatchQueue.main.async {
+                        UIApplication.shared.registerForRemoteNotifications()
+                    }
+                    
+                    StoredDefaults.standard.set(value: true, key: .didAuthorizeAPN)
                 }
                 
-                StoredDefaults.standard.set(value: true, key: .didAuthorizeAPN)
                 fulfill(true)
             }
         }
+    }
+
+    func showDistrictStatusLocalNotification(with changed: [DistrictStorageModel], observed: [ObservedDistrictStorageModel], timestamp: Int, delay: TimeInterval) {
+        let lastTimestamp: Int = StoredDefaults.standard.get(key: .districtStatusNotificationTimestamp) ?? .zero
+        guard lastTimestamp < timestamp else { return }
+        
+        let observedIds = observed.map { $0.districtId }
+        let changedObserved: [DistrictStorageModel] = changed.filter { observedIds.contains($0.id) }
+        
+        var body = ""
+        if observed.isEmpty {
+            console("📪 prepare local notification: Observed districts empty")
+            body = "DISTRICT_STATUS_CHANGE_NOTIFICATION_MESSAGE_OBSERVE_DISABLED".localized()
+        } else {
+            if changedObserved.isEmpty {
+                console("📪 prepare local notification: No changes in observed")
+                body = "DISTRICT_STATUS_CHANGE_NOTIFICATION_MESSAGE_NO_OBSERVED".localized()
+            } else if changedObserved.count == 1 {
+                console("📪 prepare local notification: 1 change in observed")
+                body = "DISTRICT_STATUS_CHANGE_NOTIFICATION_MESSAGE_OBSERVED_SINGLE".localized()
+                body.append(" \(ditrictsList(changedObserved))")
+            } else {
+                console("📪 prepare local notification: Multi change in observed")
+                body = String(format: "DISTRICT_STATUS_CHANGE_NOTIFICATION_MESSAGE_OBSERVED_MULTI".localized(), changedObserved.count)
+                body.append(" \(ditrictsList(changedObserved))")
+            }
+        }
+        
+        guard !body.isEmpty else { return }
+        showLocalNotification(title: "DISTRICT_STATUS_CHANGE_NOTIFICATION_TITLE".localized(), body: body, delay: delay)
+        
+        StoredDefaults.standard.set(value: timestamp, key: .districtStatusNotificationTimestamp)
     }
     
     func update(token: Data) {
@@ -186,6 +230,52 @@ extension NotificationManager: NotificationManagerProtocol {
         return String(data: jsonData, encoding: .utf8)
     }
     
+    func registerAPNSIfNeeded() {
+        UNUserNotificationCenter.current().getNotificationSettings { (settings) in
+            switch settings.authorizationStatus {
+            case .authorized:
+                guard StoredDefaults.standard.get(key: .didAuthorizeAPN) == nil else { return }
+                DispatchQueue.main.async {
+                    UIApplication.shared.registerForRemoteNotifications()
+                }
+                
+                StoredDefaults.standard.set(value: true, key: .didAuthorizeAPN)
+            default: ()
+            }
+        }
+    }
+    
+    private func showLocalNotification(title: String?, body: String, delay: TimeInterval) {
+        let content = UNMutableNotificationContent()
+        content.sound = UNNotificationSound.default
+        if let title = title {
+            content.title = title
+        }
+        content.body = body
+        
+        let trigger = UNTimeIntervalNotificationTrigger(timeInterval: delay, repeats: false)
+        let request = UNNotificationRequest(identifier: Constants.districtNotificationIdentifier, content: content, trigger: trigger)
+        
+        console("🚀 schedule notification with delay: \(delay)")
+        UNUserNotificationCenter.current().add(request) { error in
+            if let error = error {
+                console("😡 Local notification error \(error)", type: .error)
+            }
+        }
+    }
+    
+    private func ditrictsList(_ changedObservedDistricts: [DistrictStorageModel]) -> String {
+        var data: [String] = []
+        
+        for district in changedObservedDistricts {
+            let districtName = "\(String(format: "DISTRICT_NAME_PREFIXED".localized(), district.name)) - \(district.localizedZoneName)"
+            data.append(districtName)
+        }
+        
+        return data.joined(separator: "; ")
+    }
+    
+    
     private func subscribeTopics() {
         let didSubscribedFCMTopics: Bool = StoredDefaults.standard.get(key: .didSubscribeFCMTopics) ?? false
         guard !didSubscribedFCMTopics else {
@@ -224,12 +314,15 @@ extension NotificationManager: NotificationManagerProtocol {
 
 extension NotificationManager: UNUserNotificationCenterDelegate {
     func userNotificationCenter(_ center: UNUserNotificationCenter, willPresent notification: UNNotification, withCompletionHandler completionHandler: @escaping (UNNotificationPresentationOptions) -> Void) {
+        guard notification.request.identifier != Constants.districtNotificationIdentifier else { return }
+        
         userInfo = notification.request.content.userInfo
 
         completionHandler([.alert])
     }
     
     func userNotificationCenter(_ center: UNUserNotificationCenter, didReceive response: UNNotificationResponse, withCompletionHandler completionHandler: @escaping () -> Void) {
+        guard response.notification.request.identifier != Constants.districtNotificationIdentifier else { return }
         
         userInfo = response.notification.request.content.userInfo
         completionHandler()
@@ -239,4 +332,5 @@ extension NotificationManager: UNUserNotificationCenterDelegate {
 extension StoredDefaults.Key {
     static let didSubscribeFCMTopics = StoredDefaults.Key("didSubscribeFCMTopics")
     static let didAuthorizeAPN = StoredDefaults.Key("didAuthorizeAPN")
+    static let districtStatusNotificationTimestamp = StoredDefaults.Key("districtStatusNotificationTimestamp")
 }